added bullet obstacle detection. collision soon.

+++ src/battle/bots/game/objects/PositionedGameObject.java
@@ -6,28 +6,28 @@ import java.awt.Graphics;
 import java.awt.Rectangle;
 
 public abstract class PositionedGameObject extends GameObject {
-    private int x;
-    private int y;
+    private double x;
+    private double y;
 
-    protected PositionedGameObject(Rectangle hitbox, int x, int y) {
+    protected PositionedGameObject(Rectangle hitbox, double x, double y) {
         super(hitbox);
         this.x = x;
         this.y = y;
     }
 
-    public int getX() {
+    public double getX() {
         return this.x;
     }
 
-    public int getY() {
+    public double getY() {
         return this.y;
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
